rename trip places, voting jsx and remove combobox
