+++ client/components/PlacesForCurrentTrip.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import '../stylesheets/styles.css';
 import { Link } from 'react-router-dom';
-import Row from './Row.jsx';
+import Row from './VotingForPlaces.jsx';
 
 const Places = (props) => {
   // fetching all places for the current trip and storing them to currentPlacesInfo in state
