+++ client/components/TripDetails.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import '../stylesheets/styles.css';
 import AddTrip from './AddTrip';
-import Places from './Places.jsx';
-import Row from './Row.jsx';
+import Places from './PlacesForCurrentTrip.jsx';
+import Row from './VotingForPlaces.jsx';
 import Map from './map.jsx';
 import AddBuddy from './AddBuddy.jsx';
 import { Link, useParams } from 'react-router-dom';
