+++ client/components/VotingForPlaces.jsx
