+++ package.json
@@ -21,9 +21,7 @@
     ]
   },
   "dependencies": {
-    "@fdefelici/react-bootstrap-combobox": "^1.13.0",
     "@googlemaps/js-api-loader": "^1.14.3",
-    "@reach/combobox": "^0.17.0",
     "@react-google-maps/api": "^2.12.0",
     "@react-types/combobox": "^3.5.1",
     "@testing-library/jest-dom": "^5.16.4",
